+++ method_init/init_matrix.py
@@ -33,7 +33,7 @@ def get_voxel_coordinate(d, n):
     z = dz*np.arange(nz)
 
     # assemble the coordinate array
-    xyz = np.stack((x[idx_x], y[idx_y], z[idx_z]), axis=3)
+    xyz = np.stack((x[idx_x], y[idx_y], z[idx_z]), axis=3, dtype=np.float64)
 
     return xyz
 
@@ -59,31 +59,31 @@ def get_incidence_matrix(n):
     idx = idx_x+idx_y*nx+idx_z*nx*ny
 
     # create the sparse matrix
-    A_incidence = sps.csc_matrix((n, 3*n))
+    A_incidence = sps.csc_matrix((n, 3*n), dtype=np.int64)
 
     # assign the diagonal, each voxel is connected to three faces with positive indices
     data = np.ones(n)
     idx_row_col = np.arange(n)
-    A_incidence += sps.csc_matrix((data, (idx_row_col, 0*n+idx_row_col)), shape=(n, 3*n))
-    A_incidence += sps.csc_matrix((data, (idx_row_col, 1*n+idx_row_col)), shape=(n, 3*n))
-    A_incidence += sps.csc_matrix((data, (idx_row_col, 2*n+idx_row_col)), shape=(n, 3*n))
+    A_incidence += sps.csc_matrix((data, (idx_row_col, 0*n+idx_row_col)), shape=(n, 3*n), dtype=np.int64)
+    A_incidence += sps.csc_matrix((data, (idx_row_col, 1*n+idx_row_col)), shape=(n, 3*n), dtype=np.int64)
+    A_incidence += sps.csc_matrix((data, (idx_row_col, 2*n+idx_row_col)), shape=(n, 3*n), dtype=np.int64)
 
     # faces along x direction (faces with negative indices)
     idx_col = idx[0:-1, :, :].flatten()
     idx_row = idx[1:, :, :].flatten()
     data = -np.ones((nx-1)*ny*nz)
-    A_incidence += sps.csc_matrix((data, (idx_row, 0*n+idx_col)), shape=(n, 3*n))
+    A_incidence += sps.csc_matrix((data, (idx_row, 0*n+idx_col)), shape=(n, 3*n), dtype=np.int64)
 
     # faces along y direction (faces with negative indices)
     idx_col = idx[:, 0:-1, :].flatten()
     idx_row = idx[:, 1:, :].flatten()
     data = -np.ones(nx*(ny-1)*nz)
-    A_incidence += sps.csc_matrix((data, (idx_row, 1*n+idx_col)), shape=(n, 3*n))
+    A_incidence += sps.csc_matrix((data, (idx_row, 1*n+idx_col)), shape=(n, 3*n), dtype=np.int64)
 
     # faces along z direction (faces with negative indices)
     idx_col = idx[:, :, 0:-1].flatten()
     idx_row = idx[:, :, 1:].flatten()
     data = -np.ones(nx*ny*(nz-1))
-    A_incidence += sps.csc_matrix((data, (idx_row, 2*n+idx_col)), shape=(n, 3*n))
+    A_incidence += sps.csc_matrix((data, (idx_row, 2*n+idx_col)), shape=(n, 3*n), dtype=np.int64)
 
     return A_incidence
